updated samples to remove references to proxyUrl and customAgentOptions
