+++ samples/msal-node-samples/auth-code-distributed-cache/src/AuthProvider.ts
@@ -63,7 +63,6 @@ export class AuthProvider {
                     logLevel: LogLevel.Verbose,
                     piiLoggingEnabled: false,
                 },
-                // proxyUrl: "http://localhost:8888" // uncomment to capture traffic with Fiddler
             },
         } as Configuration;
