+++ samples/msal-node-samples/client-credentials-distributed-cache/src/AuthProvider.ts
@@ -65,7 +65,6 @@ export class AuthProvider {
                     logLevel: LogLevel.Trace,
                     piiLoggingEnabled: false,
                 },
-                // proxyUrl: "http://localhost:8888" // uncomment to capture traffic with Fiddler
             },
         } as Configuration;
