+++ samples/msal-node-samples/custom-INetworkModule/package.json
