+++ samples/msal-node-samples/custom-INetworkModule/tsconfig.json
